Refresh documentation on network connectivity
The page was not in line with the rest of documentation, had multiple bugs,
incorrect information, typos, grammar issues.

This patch reworks the structure and contents of the page from the ground up,
with the emphasis on providing timely examples and capturing just the features
that are already implemented.

Fixes #91