Fixed MAC address format in regular text
The previous attempt [1] to fix the MAC address format in the table
resulted in spurious backslashes in the result output.

This is the second attempt to make the MAC addresses look as they should
in the output. Now instead of escapes, we use ``-strings.

This new change is now validated with the actual web server running. It
works as expected.

[1] https://github.com/kubevirt/user-guide/pull/103