Add deprecation warning
A new reconciled KubeVirt documentation was released as part of the new
kubevirt.io Web site.  Since it is going to replace this User Guide in
the near future, it is fair to add deprecation warning to the User Guide
front page.

The new documentation <http://kubevirt.io/docs/> is based on the User
Guide, but the contributors worked very hard over last month to sync it
with recent KubeVirt features.  At this very moment, the new
documentation represents an up-to-date superset of the User Guide.  From
the UX perspective it is essential to convey this message and redirect
KubeVirt users to the new documentation.