Remove note about other SR-IOV device plugins
The "Intel" device plugin now supports other vendors like Mellanox and
emerges as the only solution for SR-IOV in Multus context. I don't
anticipate other plugins in the future, so it seems better to remove the
note and not confuse readers.

Any support for alternative plugins [if they emerge] will belong to
SR-IOV operator and will be abstracted out from the API user anyway.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>