ipv6, cloud-init: configure static ipv6 via networkData
Use the cloud-init networkData section to configure the IP
addresses of the VM. Network data is evaluated before userData,
and as such, when the userData script kicks in, the IP
configuration is already in place.

It is also aestethically a lot nicer to configure the IP addresses
using this declarative format instead of by running a user side
script.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>