dual stack, probes: document limitations
Document the current limitations (and work-around) of HTTP / TCP
probes on dual-stack networking configurations.

The default pod address is currently always an IPv4 address. An
issue requesting it to be configurable has been proposed in [0];
when it is fixed, the user-guide must be updated accordingly (since
it is possible that this limitation no longer applies).

[0] - https://github.com/containerd/cri/issues/1278

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>