instancetype: Update versioning section (#598)
While 0af7617caa4513fe223107dc33165e86e4ae5296 updated the format of the
data captured within a ControllerRevision it did not update the timing
of when they are actually captured.

This change updates this section now that PR #8346 [1] has landed moving
the creation of ControllerRevisions earlier in the reconcile loop.

[1] https://github.com/kubevirt/kubevirt/pull/8346

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>