Introduce horizontal menu (#650)
* Keep high-level navigation in a horizontal panel

Today, all the pages are listed in the left vertical menu. With over 60
items in this menu, it can be hard to navigate. By introducing this
high level separation, all the main sections should be right at the
top of the page.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Hide the navigation panel in one-page sections

There is no point in having a navigation panel with a single item in
it.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Move contributing section outside appendix

It was the only page in the appendix. Keeping it on the top-most level
will help people find the contribution guide.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Use bold and lighter text in horizontal menu

The original style was hard to read on the teal background.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Move release notes section behind actual docs

It is not as interesting as operations and VM docs.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Unify capitalization of text in the menu

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Hide ToC in the quickstart page

There is only a single heading, so the table of contents is not needed.

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Highlight selected horizontal menu with underline

Signed-off-by: Petr Horacek <hrck@protonmail.com>

* Keep the horizontal navigation always visible

Signed-off-by: Petr Horacek <hrck@protonmail.com>

---------

Signed-off-by: Petr Horacek <hrck@protonmail.com>