Remove Web Console documentation (#653)
The page is out of date. Keeping it in sync with the development of the
Web Console would require a lot of attention. Also, the UI should be
self-explanatory.

The second reason for this change is, that this specific UI
implementation is connected with a specific vendor. Removing it would
keep the user-guide neutral.

Signed-off-by: Petr Horacek <phoracek@redhat.com>