network: remove macvlan examples (#679)
The provided examples on the `interfaces and networks` section are using
the macvlan plugin, which we later on list as something that doesn't
work.

This commit provides a functional example to the user instead, which
will provide a better experience to the user (it works, and it is not
listed as not working later on).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>