Live Migration: Add primary pod NIC name limitation
Following kubevirt/kubevirt#13018, it is possible to have
a VM whose primary NIC is connected to a pod NIC with
a name other than `eth0`.

In order for the migration to be successful,
it is required for the both the source and target virt-launcher pods
to have the same primary NIC name, for example: `eth0`.

[1] https://github.com/kubevirt/community/pull/324

Signed-off-by: Orel Misan <omisan@redhat.com>