net, ifaces: Fix broken example
It currently says 'configured via cloud init, as shown below',
the example below has no cloud init.  Then in the bottom the note
inbdicates  'The IPv6 address for the VM and default gateway **must** be the ones
shown above', but nothing is showing above.
Hence resore the cloud-init part from the older documentation.

Signed-off-by: Nir Dothan <ndothan@redhat.com>