net, ifaces: Fix ports section (#881)
* net, ifaces: Fix ports documentation

The current documentation is incorrect. It indicates that the ports attribute
serves as some sort of firewall, when in fact is has a diffent effect on each
net binding. It is basically either a list of ports to forward or it has no effect
at all.

Signed-off-by: Nir Dothan <ndothan@redhat.com>

* net, ifaces: Fix broken example

It currently says 'configured via cloud init, as shown below',
the example below has no cloud init.  Then in the bottom the note
inbdicates  'The IPv6 address for the VM and default gateway **must** be the ones
shown above', but nothing is showing above.
Hence resore the cloud-init part from the older documentation.

Signed-off-by: Nir Dothan <ndothan@redhat.com>

---------

Signed-off-by: Nir Dothan <ndothan@redhat.com>